Set accept header according to the the produces content-types defined in the API
This fix uses the content-type defined in the operation and sets the accept header, to contain all of these content-types. If the option `responseContentType` is provided, only this value will be set as accept header.

resolves #1116
